floating

+++ 0x0D-preprocessor/1-pi.h
@@ -1,6 +1,6 @@
 #ifndef __MAIN__
 #define __MAIN__
-#define PI (long) 3.14159265359
+#define PI (float) 3.14159265359
 
 
 #endif // MAIN
